Smaller font for gmon and shell helper animated gifs
Looks like github has a max image width of 914px, above which
it'll scale the image. Which results in fuzzy letters :-(

So user a smaller font in order to make these animated gifs
narrower.

Also tweak the scriptedDemo hack for making sure state on disk
changes before either gmon or the simulated bash prompt
update. It's still a hack though.